requirements.txt
metaflow==2.4.3
pandas==1.2.4
pyarrow==12.0.0
numpy==1.24.3
comet-ml==3.21.0
python-dotenv==0.19.1
tqdm==4.62.3
wandb==0.15.4